+++ spotifywebapipython/spotifyclient.py
@@ -4357,7 +4357,7 @@ class SpotifyClient:
         This method requires the `user-read-playback-state` scope.
 
         Returns:
-            A list of `Device` objects that contain the device details.
+            A list of `Device` objects that contain the device details, sorted by name.
                 
         Raises:
             SpotifyWebApiError: 
@@ -4391,7 +4391,11 @@ class SpotifyClient:
             items = msg.ResponseData.get('devices',[])
             for item in items:
                 result.append(Device(root=item))
-        
+
+            # sort items on Name property, ascending order.
+            if len(result) > 0:
+                result.sort(key=lambda x: (x.Name or "").lower(), reverse=False)
+       
             # trace.
             _logsi.LogArray(SILevel.Verbose, TRACE_METHOD_RESULT_TYPE % (apiMethodName, 'list[Device]'), result)
             return result
@@ -4411,6 +4415,7 @@ class SpotifyClient:
 
     def GetPlayerNowPlaying(self, 
                             market:str=None, 
+                            additionalTypes:str=None
                             ) -> PlayerPlayState:
         """
         Get the object currently being played on the user's Spotify account.
@@ -4426,6 +4431,13 @@ class SpotifyClient:
                 Note: If neither market or user country are provided, the content is considered unavailable for the client.  
                 Users can view the country that is associated with their account in the account settings.  
                 Example: `ES`
+            additionalTypes (str):
+                A comma-separated list of item types that your client supports besides the default track type.  
+                Valid types are: `track` and `episode`.  
+                Specify `episode` to get podcast track information.  
+                Note: This parameter was introduced to allow existing clients to maintain their current behaviour 
+                and might be deprecated in the future. In addition to providing this parameter, make sure that your client 
+                properly handles cases of new types in the future by checking against the type field of each object.
                 
         Returns:
             A `PlayerPlayState` object that contains the currently playing media, or null if nothing is playing.
@@ -4459,6 +4471,8 @@ class SpotifyClient:
             urlParms:dict = {}
             if market is not None:
                 urlParms['market'] = market
+            if additionalTypes is not None:
+                urlParms['additional_types'] = additionalTypes
 
             # execute spotify web api request.
             msg:SpotifyApiMessage = SpotifyApiMessage(apiMethodName, '/me/player/currently-playing')
@@ -4488,6 +4502,7 @@ class SpotifyClient:
 
     def GetPlayerPlaybackState(self, 
                                market:str=None, 
+                               additionalTypes:str=None
                                ) -> PlayerPlayState:
         """
         Get information about the user's current playback state, including track or episode, progress, 
@@ -4504,6 +4519,13 @@ class SpotifyClient:
                 Note: If neither market or user country are provided, the content is considered unavailable for the client.  
                 Users can view the country that is associated with their account in the account settings.  
                 Example: `ES`
+            additionalTypes (str):
+                A comma-separated list of item types that your client supports besides the default track type.  
+                Valid types are: `track` and `episode`.  
+                Specify `episode` to get podcast track information.  
+                Note: This parameter was introduced to allow existing clients to maintain their current behaviour 
+                and might be deprecated in the future. In addition to providing this parameter, make sure that your client 
+                properly handles cases of new types in the future by checking against the type field of each object.
                 
         Returns:
             A `PlayerPlayState` object that contains the currently playing media, or null if nothing is playing.
@@ -4537,6 +4559,8 @@ class SpotifyClient:
             urlParms:dict = {}
             if market is not None:
                 urlParms['market'] = market
+            if additionalTypes is not None:
+                urlParms['additional_types'] = additionalTypes
 
             # execute spotify web api request.
             msg:SpotifyApiMessage = SpotifyApiMessage(apiMethodName, '/me/player')